Add initial support for stream upstream metrics
This includes only the subset of per-upstream/peer metrics that already
exist for non-stream upstreams which are applicable to stream/UDP
upstreams. There are additional stream/UDP-specific metrics that should
be added later, but this is a useful start.

Note: This propagates the existing incorrect help text for the "fails"
metric (it shouldn't be "Active connections") because I'm not sure what
the correct help text should be.